moved Dagostini implementation to examples
